restructured; added sample inputs
